Move status update (#78)
* update status after installation

* move instance update funcs

* split update state fns into smalled pieces
