+++ components/operator/internal/state/apply.go
@@ -13,7 +13,6 @@ import (
 func sFnApplyResources(_ context.Context, r *reconciler, s *systemState) (stateFn, *ctrl.Result, error) {
 	// set condition Installed if it does not exist
 	if !s.instance.IsCondition(v1alpha1.ConditionTypeInstalled) {
-		s.setState(v1alpha1.StateProcessing)
 		s.instance.UpdateConditionUnknown(v1alpha1.ConditionTypeInstalled, v1alpha1.ConditionReasonInstallation,
 			"Installing for configuration")
 	}
